10. Creation <Road To Complete - 9°, 10°>
In this commit I added the FPS weapon to the player. Added the prefabs in the scene with colliders so we can interact with them and pick them up. The weapons of the player will be set as active false to become active true once the player picks up. And add the lights of indicators and the scary light of the dark room.